K8S-423: Correct Pod Environment Type
The underlying type for setting pod environment variables is EnvVar
which is a struct and not the map our documentation alluded to.
Corrects the documentation, adds in validation and a few tests to
confirm.  Also fixes a few tests which missed the V1 update.  And also
makes the suite report a fail if decode failed unexpectedly, as the
prior fixes were not picked up in CI.

Change-Id: Ia74f0ba6590e222968504a6dfecc2ea88db19d5d
Reviewed-on: http://review.couchbase.org/96380
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>