K8S-433: cancel failed add back
If a node has been failed over and an attempt to add it back via delta recovery
also fails then we should try to add it back using a full recovery.

Change also support capturing of errors from retry loop in order to detect
the type of error that has occurred during a rebalance (or any other retry call).

Change-Id: Ib345ab910d5c5dadc1ca101e3bb1324a6fc3f559
Reviewed-on: http://review.couchbase.org/96639
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>