K8S-266: Don't set a security context by default
This was initially done to solve an issue of running containers on
OpenShift, but it didn't actually work. OpenShift randomizes the
user that starts a particular container so we can't set the user
and group the container should be run with because if it doesn't
match what OpenShift picks the container won't start properly.

Change-Id: Icb33290c77129413d4b57308dda67a9fdcd2d83f
Reviewed-on: http://review.couchbase.org/96840
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Mike Wiederhold <mike@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>