K8S-435: Filter Exposed Ports
When exposing service ports we expose the entire set and not those
relevant to the services running on the actual node.  This filters the
set applied to the node based on what is defined in the server
configuration.

Sadly there was an inconsistency between the specification and the
actual services that were created making this difficult.  As such I've
enumerated the services in the specification which makes all things
less error prone as we have type safety and not just a random string.
Test is the one dependency which still uses strings sadly, but I have a
plan...

Change-Id: I3685fc6c33938fdbc8f4dce862802fd691fec913
Reviewed-on: http://review.couchbase.org/96861
Reviewed-by: Korrigan Clark <korrigan.clark@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>