K8S-469: Init Node Services Before Addition
We previously added node services after balancing in as the reconcile
FSM didn't check for things like services, however you may be waiting
hours for a re-balance to complete.  If we move to before the re-balance
then at least the XDCR agents can start streaming some data in the
periods where the vbucket maps are mostly stable.

As a result the FSM init function needs to see whether any services need
to be reconciled before allowing execution.  The same applies to server
groups.  It is accepted that rather than duplicating the core
functionality of these reconcile functions we should just let the FSM
execute regardless, but given the time we'll accept the pain and clean
up once the next release goes out the door.

Change-Id: I6e635973e0ef05a50946409fbf4fff0c23132fa7
Reviewed-on: http://review.couchbase.org/97137
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>