K8S-445: Recommend Failover when missing volumes
It's possible for Persistent Volume Claims(s) associated with a Pod to be deleted
or to be unbound from it's underlying Volume.  Depending on the service the volume
belongs to, the Couchbase Server may continue to 'function' as healthy when in
really the data is not being persisted.  This change detects that there is actually
a problem with the underlying volumes and recommends the user to failover the node.

After user performs manual failover, the operator will then remove the node from cluster and create
a new node.

NOTE: Pod with unhealthy volume aren't automatically deleted by operator because this could lead
to unexpected data loss when failover cannot be performed.

Change-Id: I2d8cebe73f871be23e34792d51695c4e1f9d00d5
Reviewed-on: http://review.couchbase.org/97411
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>