K8S-479: Check Rebalance Status
When we poll the cluster status post rebalance we should also check to
see if the cluster thinks a rebalance is still necessary.  This catches
situations where nodes have been correctly balanced in/out however a
subordinate service (eventing) has raised an error.

So on to the actual test... first we have to wait for the deployment to
warm up before we allow rebalance for determinism.  Next the insert
routine didn't actually work as the break was only scoped to the select
statement, it also failed pretty epically when raising Fatal so I've
made this better.  The cluster scale was pretty much untestable so I've
made that smaller.  The resizes didn't set the correct value implying
this hasn't actually been tested.  We should also wait for the insert
routine to exit so we know it does, and so we know it's dead avoiding
races.

Change-Id: I31f2693752b4e799df298c2448108fa5798c2eb9
Reviewed-on: http://review.couchbase.org/97376
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>