K8S-500: Improve Quick Starts
We do a bad job of disabiguating the names of various resources in the
quick start guides and also expect people to use the names we give them
which caused confusion when they are altered.  I've reworked the
examples to set parameters via environment variables so you only need to
change a single value, the names are very clear about what is being
changed.  Each of these has been documented to explain what it is being
used for, and whether it can be changed or not (e.g. is hard coded in
some web resource).

Also sneaking in some more examples for OpenShift log collection.

Change-Id: Id48f52af9d08524f6854017e57a72f9e65b9684b
Reviewed-on: http://review.couchbase.org/97638
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>