K8S-497: Failing Autofailover
NS server is preventing the auto failover from occurring as it sees two
nodes as down (possibly due to one of the new nodes still transitioning
from failedAdd to pendingAdd.

The test doesn't actually work, so I've fixed it, removed all the
channel code as it didn't have any timeouts associated with it, so
basically it would just hang indefinitely and made the failover reset
retry as it's entirely possible to try it on a dead pod before DNS is
updated.

Change-Id: I1c23303f1eb178101d16598412733980190cfdbf
Reviewed-on: http://review.couchbase.org/97778
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ashwin <ashwintrojan+github@gmail.com>
Tested-by: Ashwin <ashwintrojan+github@gmail.com>