K8S-510: fix segfault when removing Server Configs
Fix is simply a nil check on the members ServerConfig.

When Server Config is removed from Spec it's Pod is deleted.
We then try to get the Pod's ServerConfig to determine if
removed Pod has volumes that should be deleted.  However,
Server Config has also been removed and is nil.

Change-Id: Iadac68c40834682c51feb772846c9dd1673b820e
Reviewed-on: http://review.couchbase.org/97875
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>