K8S-514: Ports Must Handle Missing Server Classes
We handle the case where a pod is gone in the port filter but not when
server classes are deleted and the pod is about to be.  Some basic
spring cleaning to error handling and then a switch to process expected
error types.  In this specific case we leave the ports alone so client
services continue to work.

Obviously as a result you cannot change the exposed features at the same
time as we can't filter properly without interrogating the underlying
pod's labels or Couchbase server, both of which don't fit with the
current model.

Change-Id: I622e38052339141bda084374bde26a220386e443
Reviewed-on: http://review.couchbase.org/97921
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>