K8S-518: Fix Volume Check Race
If a server class is removed after the main reconcile fsm has run but
before we check the volumes we will erroneously raise a volume unhealthy
event.  The simple fix is to ignore missing server configurations,
relying on the fact the pod will be deleted next time around.

The longer term fix will be to make CRD updates atomic as regards the
reconcile loop (e.g. cache and only make active at the beginning of a
reconcile), that way we can guarantee that the pods in the missing
server class will be removed before the volume check occurs.

Change-Id: I95c794dd4d7964ce84392e3c9a530f61fbd9be9e
Reviewed-on: http://review.couchbase.org/98066
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>