K8S-545: Fix node recovery after failed add back node
If a node goes down while being delta recovered we should finish reconcile loop rather than continuing and allowing node to be deleted since this will also delete it's volumes and logs.

On 2nd pass of reconcile we will get the current state of the node and if it is actually down the node can be properly recovered so that it's logs are retained.

Change-Id: I8c04e5a58bc3c5d09a0f7384312cae7987368836
Reviewed-on: http://review.couchbase.org/98459
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>