K8S-542: Only skip recovery of a down node if it's pod is over 60 seconds old
It's possible to have a running pod that contains a crashed or frozen
Couchbase Server instance. If Couchbase Server sees that node as down,
but it's pod is still running then we need to consider restarting the
pod.

It should also be pointed out that a node marked as down may have been
recently restarted and is in the process of joining the cluster so
that it can be delta recovered later so we can't assume that a node
reported as down by Couchbase with a running pod should be restarted
every time.

To address this we will not restart a down node if it's pod is less than
60 seconds old in order to give it time to join into the cluster. If the
pod is still down after 60 seconds we will consider it actually down and
allow it to be restarted.

Change-Id: Ia99810d56acd422297cf16ebaf2785a3ab715e39
Reviewed-on: http://review.couchbase.org/98531
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>