K8S-549: Dynamic Server Groups
The core problem here is that the framework decided to clean all zone
labels, which in turn broke all kinds of cloud storage as these are
required to match those attached to volumes.  Remove the deletion.

Second problem was that the RZA tests totally ignored the configuration
and used what ever cluster it saw fit, which in this case certainly
wasn't my kops cluster in AWS!

Using my cluster we don't need to use the decorator as the labels are
already there, with the magical QE ones they can still use it.  When
each test now requests its list of availability zones, it does so via
the API, which works in all cases.

TestRzaServerGroupDown doesn't work as that tries to talk directly to
server rather than look at the CRD status.  In this case only the API is
exposed via an ELB instance, so trying to change the port to the admin
port isn't going to work without a LoadBalancer service.

TestRzaAntiAffinityOn didn't work as it assumed that the number of
schedulable nodes was equal to the number of server groups which won't
always hold.

Change-Id: I5e57bb828609682f06d3c245cee24c04cd5bc29c
Reviewed-on: http://review.couchbase.org/98695
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>