K8S-572: Support Environment Vars in cbopinfo
Most commands allow KUBECONFIG to be specified in the environment so we
should support this too.  Rather that treating one parameter as a
special flower we allow all to be specified by environment much like
OpenStack commands for those used to such behaviour.

As an aside cbflag doesn't support non-flag arguments which aren't
commands so it can't be used.

Change-Id: Ie81d62fcfad816877925857c79026c640c6d8e33
Reviewed-on: http://review.couchbase.org/99710
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>