K8S-584: Fix Runloop DoS Attack
The main purpose of this change is to avoid a situation where we receive
constant CR updates and don't ever reconcile.  It will happen 8 seconds
after the last run regardless of any updates that are queued up so we
can at least patch up as best we can in the face of adversity.  This is
kind of controvertial so I'm open to suggestions as to whether things
were better before or not.

While here there are a number of cleanups.  The main body is moved into
a function, meaning timing stats and status updates can be done in
deferred function calls saving code repetition.  Also I noted that the
PV recovery code would get implicitly run we just let updateMembers()
get run regardless of whether there are running pods or not.

Change-Id: Ib84556749762c96b8017d44bc9d28eba6859f517
Reviewed-on: http://review.couchbase.org/99798
Tested-by: Simon Murray <simon.murray@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>