K8S-599: Fix updateMembers to include all persisted members
Removes code that only sync'd memberset to nodes that were
in specfic cluster states (Active, Pending, Failed, Down).
This is problematic for 2 reasons when operator is recovering:
1. A node in Active/Pending state may actually be foreign,
   therefore it should not be tracked as a member.
2. Node may be in a state we haven't accounted for but is a
   legitmate member.  And when we miss a node it gets marked
   as unmanaged and removed (K8S-599)

The fix here is to rebuild the members strictly from Running Pods
and Persistent Volumes.  Any nodes which couchbase knows about
outside of these require manual failover since they can only
be of the following types:
  //  1. A foreign node
  //  2. A node with both deleted Pod and Volume
  //  2. An ephemeral node with deleted Pod.

Change-Id: I40f5d554c6661e5e8eda63834a3ad372f3f2189d
Reviewed-on: http://review.couchbase.org/99999
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>