K8S-504: PVC Janitor Process
Spawns a per-cluster process which is responsible for cleaning up old
logging volumes.  It is driven by two new CR attributes which control
how long to retain for and how many to retain.  The janitor process
first looks for orphaned volumes and marks them as detached with a time
stamp annotation.  This is then used to remove old logs that have been
around too long, or the oldest if there are too many.

Change-Id: I68562d01da999621cd6c2abf0a815e41b5c4cbec
Reviewed-on: http://review.couchbase.org/100124
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>