K8S-610: Log Volume Collection
Automation for collection of logs from ephemeral server classes.  It
first looks for detached server logs, sorts them by time then prints out
a table of available logs.  The user is then prompted to select which to
collect.  In parallel it will then spawn an ephemeral pod with the logs
mounted, run cbcollect_info and download the logs locally.

I think in future we should have a --redact-server-logs option to select
which logs to download as space may become a concern.  Secondly I think
we should combine log volumes and running pods into a table and let the
user select from running and dead instances what to collect, makes the
whole experience better and consistent.  But not today... 1.2.0.

Change-Id: I9a9816c550857005e22cf223ace5d0d03c738ea1
Reviewed-on: http://review.couchbase.org/100240
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>