K8S-613: Supportability Validation
Added in support ability constraints.  This enforces that if one class
is supportable they all are.  It also enforces that stateful services
e.g. data, index & analytics, have a default mount.  Finally when
mounting claims we ensure the service is actually defined for this
class.

Test 0074 is now invalid as specifying a data mount with a logs mount
will ensure that the data service is enabled, and also enforce that to
be supported the data service needs a default mount.

Change-Id: Ie23eb11a4c6ea9ee91bf1902ff671032eecdac3a
Reviewed-on: http://review.couchbase.org/100277
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>