K8S-24: Upgrade
Strip out the old upgrade code.  This new version is decidedly more
stateless, relying on pod annotations.  Upgrades only work after the
cluster is reconciled back into a safe state.  It selects the lowest pod
for upgrade and does a swap rebalance, simple!

The core of the code is fairly unconstrained which makes it very
flexible.  Things like protecting which upgrades and roll backs can be
performed are delegated to the admission controller.

Change-Id: Ia0931fe1a2d67351807d53a6508161f9707be7ef
Reviewed-on: http://review.couchbase.org/100631
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>