K8S-474: Unify Log Collection
At present Server logs are all collected and left on the running pods.
For pvc logs these are interactively selected and downloaded.  This puts
all the options in a table, allows simple collection of one, many or all
logs, prompts for log redaction to save space, and also removes logs
from ephemeral storage.

The CLI is extended to allow QE to poll for log sources and to
explicitly specify which to gather and whether to apply redaction.

Change-Id: I8c13505f1437504da86e17055c1e384969564c90
Reviewed-on: http://review.couchbase.org/100713
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>