K8S-648: Unstructed Validation
We do the CRD validation too late, after decoding.  If we did it on
unstructued data before decoding we'd get the same output as from the
API however we just get gobbledegook from the json deserializer.

Aside from that I've cleaned up the main function which allows tests to
be run individually with a -run flag.  The value of `path` changed to
mean the incoming CR or the existing CR depending on the test context,
so I've simplified this so `path` is always the incoming CR.  The
loading of YAML now may catch CRD validation errors, which are now
checked against the expected errors test parameter.

Change-Id: Ic5cf7b412084b6c24f576c685da135b60295a436
Reviewed-on: http://review.couchbase.org/100892
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>