K8S-413: Generic Object Updates and Verification
I discovered that trying to change the CRD in any non-trivial way
causes a lot of work in test.  It dawned on me that by using
reflection we break the ties to out structured types, all update and
verfication tasks can be acheived with just two functions.

The object utility module allows generic deep value lookups via a string
path based on JSON pointers, but restricted to Go like semantics.
From this we can update or verify any value in any data struture.

Change-Id: Icd9737811aa4c2e03aa35ba3daf40401f9b0e937
Reviewed-on: http://review.couchbase.org/98004
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>