K8S-672: Use NodeServices Endpoint for Services
We used /pools/default to get alternate addresses however this is
subject to some form of filtering, /pools/default/nodeServices doesn't
suffer from this issue.

This unmasks the previously filtered ports, also performs a small
refactor as K8S-618 isn't in yet and I've already submitted the change
to gocbmgr.

Change-Id: I889cafd55a58ca0f1407dc73df8560744b057d87
Reviewed-on: http://review.couchbase.org/101800
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>