K8S-675: TLS Validation
Users usually tend to get in a pickle with TLS and expect us to debug it
for them, which we cannot as the secrets are redacted for security
reasons.  This automates the process by explicitly telling them what
they did wrong to avoid support queries :D

The code will be shared with the TLS rotation stuff so at least we can
spot silly errors in the logs before potentially installing the
certificates and breaking cluster communication.

Change-Id: I8d5415389ec40a84dd20ab6a4dc9bfd14986250b
Reviewed-on: http://review.couchbase.org/101957
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>