K8S-674: Operator Upgrades
As is always inevitable we'll make some changes upstream that are not
compatible with clusters deployed with an earlier release.  We should
spot when taking over a pre-exsting cluster and perform any necessary
upgrades to the kubernetes resources before allowing the operator to run
fully.

This version independently upgrades each resource type based on operator
version, which is now added to upgraded resources so we know what and
what not to apply in future.

Change-Id: I6563a8d91648c6c079f3eed935e095bb1620c280
Reviewed-on: http://review.couchbase.org/103290
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>