K8S-680: TLS Rotation Across Restart
If we restart the operator during a TLS rotate it will bomb out because
the code that checks cluster membership tries to contact NS server with
a new CA, and it can't, as server is using the old certificate chain.
We need to perform the rotation once we know we are doing a restart for
any detected pods i.e. not vetted against server.

Change-Id: I4b9f562ac4237e161881835a2c1e5cd0a1c014fd
Reviewed-on: http://review.couchbase.org/102094
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>