K8S-683: Fix QE Supportability
Fixes log collection for when a test unexpectedly fails.  Improves the
server logs detection code to be less reliant on command output.
Changes server log collection tests to be non-blocking (not crash).

The --collectinfo flag test cannot possibly work now, so it has been
consigned to the bin.

The log redaction stuff again cannot possibly be tested the way it was
so now does what it can.

Change-Id: I20811bdbe31a161316c60faf3e477f97fb87b0ac
Reviewed-on: http://review.couchbase.org/102200
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>