K8S-686: Update Dist Make Target
Adds in a dist build for the admission controller, with the necessary
Docker files renamed to Dockerfile and Dockerfile.rhel respectively to
keep things similar to the operator build.

Couple cleanup operations bump the version to 1.2.0 by default and also
refactored the existing artifacts build to use the environment variable.

Change-Id: I71f36ec50aa6d03577eb877f4ecf9ffd0c84956f
Reviewed-on: http://review.couchbase.org/102258
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Kenneth Lareau <kenneth.lareau@couchbase.com>