K8S-692: Improve QE CRD Validation
General clarifications of the data structures, and the main algorithm
which was hard to read.  The main chunk of the work ports the old field
mutation stuff to use JSON patches instead.  Combining with removal of
redundant stuff there's a nice saving of 1000 lines :D  Also there were
a bunch of tests (commented out) that would never work without
manipulating JSON directly, so I've deleted them.

Change-Id: Iadf68e9e615811bc42b58f9ed26319363f937aea
Reviewed-on: http://review.couchbase.org/102425
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>