K8S-698: Unify CRD Validation Naming
Removes a hack I added to clean up the test output.  Now allows you to
find the exact failling test from the console output.  Also standardizes
the test names.

While in the area it makes sense to save 6 hours of time by not
requiring the operator to be running (doesn't need to be), which also
allows the test to be run anywhere (by development!) without the need to
conjure up some non-standard zone labels.

Finally squashed a few bugs, because we can actually test this now!

Change-Id: I823abfabb79cf0219981233b86c3be7d0486930c
Reviewed-on: http://review.couchbase.org/102680
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>