K8S-701: Stateful Upgrade Tests
Adds in tests to check upgrades succeed for various death scenarios
during upgrade of a supportable cluster.

Couple enhancements here, first the order of upgrades has been made
deterministic in order to be able to kill the pod we want to.  Secondly
the pod killing code always removed the pvcs, so I've made this optional
for these tests.  The others need to be reviewed because I suspect they
aren't testing what you think they are!

Change-Id: Ib4383fb02cfc0fcc0d20f68e05dd45009de32e14
Reviewed-on: http://review.couchbase.org/102730
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>