Reduce length of PVC name
So that we don't easily hit the 63 character limit!

When clusters names are more verbose (as is the case with helm)
then members names become longer, which in turn lead to longer
names for pvc volumes...ie:
	pvc-couchbase-messy-dog-couchbase-cluster-0000-00-default

This change drops the leading 'pvc-<claim-name>' prefix leaving
only the unique elements which vary per member & cluster.

Change-Id: Ib62008e2909ab07fd8be61d9a543f11c87a71f17
Reviewed-on: http://review.couchbase.org/103162
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>