K8S-716: How To Pull Images Properly
Previously image pulling relied on SSH/Ansible to get images downloaded
and retagged.  We shouldn't rely on SSH access as it's not a given.

This instead allows the use of named images pulled directly from our
internal private repositories.  To faciliate this QE needs to set the
operator-image and admission-controller-image variables in the test
configuration (test_config.yaml) using exactly what they expect e.g.
couchbase/couchbase-operator-internal:1.2.0-827.

Authentication is again set in the same file with the docker-server,
docker-username and docker-password variables which allows access to the
internal repos.  If specified this will automatically attach pull
secrets to the operator and admission controller deployments.

Because we use explicitly tagged build versions there is NO ambiguity in
which image we are actually using.  Win, win.

Change-Id: I3d69cb819cff8efb618c6584b826aeb1bcf7362c
Reviewed-on: http://review.couchbase.org/103284
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Arunkumar Senthilnathan <arun.couchbase@yahoo.com>