K8S-733: QE Selectable Contexts
Dev tend to use explicit configuration references as the standard QE
code will try to run SSH against our cloud clusters and fail horribly
when using implicit references.  Most cloud tools will install a new
context in ~/.kube/config when dynamically provisioning, so it'd be nice
to select explicitly which context we wish to use for which cluster.

Change-Id: I47cde8ed829c90faa6e89d32c3e9a614ac44c62b
Reviewed-on: http://review.couchbase.org/103440
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>