K8S-739: Parameterize Upgrade Versions
Add in runtime configuration of an upgrade version.  Update the tests to
skip if an upgrade version has not been specified.  Update all
references to the source and target versions to be those supplied in the
configuration.

Also took the liberty of reformatting the configuration so we are not
continually reformatting comments at the end of lines, move them to the
begining of their own line, problem solved permanenetly.

Change-Id: I9f176f1b5870826b629c50abef2024691f42c19e
Reviewed-on: http://review.couchbase.org/103485
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>