K8S-742: Retry Rebalance Status
This is the major thorn in the side of QE getting perfect test scores.
Sometimes Server returns rebalance as incomplete, when next time around
it says it is complete.  We raise the "wrong" event and the tests fail.
Retry the status check if it reports as incomplete to be really certain.

Change-Id: I5a383b6f5af9847b875d864cc1e5eededaa72601
Reviewed-on: http://review.couchbase.org/103544
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>