K8S-746: Improve QE cbopinfo Code
There is all kinds of hackery involved with handling arguments that can
be cleaned up if we abstract the data type as a map and ignore empty
values.  The main goal here is to provide a default set up that
configures the kubernetes configuration, context (optional) and
namespace, so we can make these tests work in the cloud regardless of
configuration method.

Change-Id: I6c0da5f4451a2919663b621bfd3cc89948de730a
Reviewed-on: http://review.couchbase.org/103592
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>