K8S-761: Conditional PVC Tests
At present a lot of the PVC and supportability tests are untestable in a
multi-zone cluster as PVCs for a single pod may get scheduled in
us-east-1a, us-east-1b and us-east-1c.  So rather than just failing all
over the show, just ignore the test thus removing a lot of the false
positives.

Change-Id: I47078f87fb3fdb92511fa113efead61afc66328d
Reviewed-on: http://review.couchbase.org/103768
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>