K8S-774: Fix Scheduling Functions
A recent change to the scheduling code (correctly) made the assumption
that master nodes aren't schedulable rather than blindly just using the
raw number returned by the API.  However QE tend to untaint all nodes
which means this assumption didn't hold.  Instead if we check for the
NoSchedule taint as a filter it will hold in all cases.

Change-Id: I1944f392d7062124e5f8b22b4923cd2cfa43645f
Reviewed-on: http://review.couchbase.org/103965
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>