K8S-770: Fix Test Interference
The current test framework assumes persistent volumes will be reclaimed
instantly, which is not the case, and a PVC from one test can be seen in
another which triggers failures.  So two fixes are required, first that
the supportability tests correctly filter PVCs as they are cluster
scoped, and second that (like pods) we actually wait on test completion
for the PVCs to be deleted so they cannot bleed into subsequent tests.

Change-Id: I5690d8a5947bb35d217dfbaec0fac1662b287f8d
Reviewed-on: http://review.couchbase.org/104013
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>