K8S-787: Fix Resource Annotations
Couple fixes to the upgrade code.  We need to start adding version
information to all resources or we risk running upgrade tasks
erroneously.  Also a little fix to the error code for missing labels, I
thought errors were interfaces and therefore implementations needed to
be pointers, but alas no, so a bug crept in.

Change-Id: I3ad47d57bd7712efcb11d32cfd2a71938bda41dc
Reviewed-on: http://review.couchbase.org/104130
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>