K8S-794: Fix TestDiskFailureAutoFailover
Saw a failure on GKE last night for the first time.  I suspected that a
failing exec was to blame, but apparently that's a false positive.
Still gave me a reason to move the code to where it should have been to
start with and clean it up.  Have extended the timeouts as the output
suggests the tests would have completed, perhaps with different events
registered due to a rebalance problem.  Also the test itself was a bit
of a mess so I've cleaned that up.

Change-Id: I552d3ddd0f87758ab4e332a2ad5668b835717e16
Reviewed-on: http://review.couchbase.org/104252
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>