K8S-799: P1 Failures
Base image and version tests looked to be all the way broken, so I've
made them spin up a cluster, not waiting, ensuring a pod comes up but
unable to pull the image, then eventually the cluster fails.  The
unschedulable test used a broken methodology for determining memory
utilisation so that's been moved to the fixed version.  The pod killing
test during scale down was non-deterministic so we are now using more
accurating killing functions.  Final thing is to reduce a load more
repetition from the tests as it was hurting my eyes.

Change-Id: I9d70212ec5263fc5ae289e185c16c068eed9a47d
Reviewed-on: http://review.couchbase.org/104332
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>