K8S-804: More Aggressive PVC Deletion
We are seeing a 20 minute hang between tests some times implying that
something is hanging, probably the PVC code with 120 (x10 seconds)
retries.  This makes it much more agressive, deleting all PVCs in the
cluster maked as couchbase, and retrying the delete if they don't
disappear.  Also sneaking in a fix as go managed to cast 10 (number)
into time.Duration (10ns) without raising an error.

Change-Id: Ibb881df05be263f3c484564ff64e9ab3040b9765
Reviewed-on: http://review.couchbase.org/104449
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>