Helm refactoring operator chart
Removed nameOverride because it can't be applied in a consistent way
to both operators since we'll end up with dupes.

Any type in values with a create attribute will use the name "default"
when create is false unless the name field is set as an override.
This will keep override methodology consistent.

Some minor fixes in.

Change-Id: I636ea032030bcb3c933051f0be598f4656272672
Reviewed-on: http://review.couchbase.org/104615
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>