K8S-814: Another Collection of P0/P1 Fixes
Seemingly 2 minutes isn't enough to reliably propagate documents via
XDCR so up this to ten minutes.  This should rule out timing issues and
firmly point the finger at us or XDCR.

The eventing and analytics functions are all broken and may end in a
deadlock.  If an insertion failed for example it would bomb out of the
loop and wait trying to send a failure.  In the main routine it would
wait trying to deliver on the cancel channel.  I've made these work,
rerty on failure and also by using the magic of zero based indices make
the code less confusing.

The analytics tests also committed the cardinal sin of calling t.Fatal
in a go routine, which results in a crash.

Change-Id: I169833fb75140f5ce318ea5cbc48e7c2a24059b0
Reviewed-on: http://review.couchbase.org/104752
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>