K8S-817: P1 Failures VI
Rotate the pod we are talking to while inserting docs into clusters
whose topology is changing, it may fail once in a while (silently) but
not constantly which we are trying to avoid.

This highlighted another race condition where a routine failed (fix me)
after the test had finished.  We need to cleanly shut down the insertion
routines in the event of a failure with deferred execution.

Caught a bug that has been around for literally years where if the
master pod is killed before the ready/unready lists are updated we'll
get stuck forever trying to reconcile the cluster settings #omg!

Change-Id: I53fd0e87d576bb11a2f294ab569267ba80ea0f9e
Reviewed-on: http://review.couchbase.org/104828
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>