K8S-834: Deduplicate Build
It has come to my attention that we don't have a canonical source for
the licence and readme, prefering 3 copies of exactly the same thing,
which leads to one being updated and the rest forgotten about.

Sneaking in a timeout which I suspect has to do with the target cluster
taking a while to pull a Couchbase Server image.

Change-Id: I89c5048005d45feab612347cc379df0f663c0950
Reviewed-on: http://review.couchbase.org/105055
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>