K8S-844: P0 Fixes
A couple functions appear to need retries to ensure success more
frequently.  There is also a race when inserting data and balancing out
nodes as there's a possibility that the last insertion is to the
balanced out node, which fails, so we need to add a little sleep in
there.

Change-Id: If2088fb85881a745418b3cb89452167c17949c2c
Reviewed-on: http://review.couchbase.org/105160
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>