K8S-847: Generic Resources
We have 4 copies of all the manifests, one for dev, one for QE and two
for docs.  Lets get on the same page and use the same ones everywhere.
QE are confident that what is in the packages is tested, everyone is
confident that dev have done the updates.  And the best things of all
deployment gets a million times simpler.

Change-Id: Ie31c2a4f291be16424f912671c2da11559cb8d3b
Reviewed-on: http://review.couchbase.org/105232
Reviewed-by: Eric Schneider <eric.schneider@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>