K8S-865: Cull Direct Client Use
In the old world we'd directly connect to a pod via the admin service.
In the new world we randomly open a port forward to pod to simulate the
same behaviour.  Reusing old clients was okay back then, but now we need
to reinitialize on each retry.  This kills off a few more uses of client
directly to avoid them going stale, the end goal to eventually not allow
tests access to avoid the problem altogether.

Change-Id: Ib3a10da5fe5054e3eef9ff2805bea5adb863e664
Reviewed-on: http://review.couchbase.org/105539
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>